[inetstack] Enhancement: Create L4 module
