+++ src/rust/inetstack/protocols/layer4/tcp/active_open.rs
@@ -16,7 +16,7 @@ use crate::{
             PacketBuf,
             SharedLayer3Endpoint,
         },
-        tcp::{
+        layer4::tcp::{
             constants::{
                 FALLBACK_MSS,
                 MAX_WINDOW_SCALE,
