+++ src/rust/inetstack/protocols/layer4/tcp/established/background/acknowledger.rs
@@ -3,7 +3,7 @@
 
 use crate::{
     collections::async_value::SharedAsyncValue,
-    inetstack::protocols::tcp::established::background::SharedControlBlock,
+    inetstack::protocols::layer4::tcp::established::background::SharedControlBlock,
     runtime::fail::Fail,
 };
 use ::futures::never::Never;
