+++ src/rust/inetstack/protocols/layer4/tcp/established/background/retransmitter.rs
@@ -3,7 +3,7 @@
 
 use crate::{
     collections::async_value::SharedAsyncValue,
-    inetstack::protocols::tcp::established::ctrlblk::SharedControlBlock,
+    inetstack::protocols::layer4::tcp::established::ctrlblk::SharedControlBlock,
     runtime::{
         conditional_yield_until,
         fail::Fail,
