+++ src/rust/inetstack/protocols/layer4/tcp/established/congestion_control/cubic.rs
@@ -15,7 +15,7 @@
 
 use crate::{
     collections::async_value::SharedAsyncValue,
-    inetstack::protocols::tcp::{
+    inetstack::protocols::layer4::tcp::{
         established::congestion_control::{
             CongestionControl,
             FastRetransmitRecovery,
