+++ src/rust/inetstack/protocols/layer4/tcp/established/mod.rs
@@ -11,7 +11,7 @@ use crate::{
     collections::async_queue::SharedAsyncQueue,
     inetstack::protocols::{
         layer3::SharedLayer3Endpoint,
-        tcp::{
+        layer4::tcp::{
             congestion_control::CongestionControlConstructor,
             established::ctrlblk::SharedControlBlock,
             segment::TcpHeader,
