+++ src/rust/inetstack/protocols/layer4/tcp/passive_open.rs
@@ -19,7 +19,7 @@ use crate::{
             PacketBuf,
             SharedLayer3Endpoint,
         },
-        tcp::{
+        layer4::tcp::{
             constants::FALLBACK_MSS,
             established::{
                 congestion_control::{
