+++ src/rust/inetstack/protocols/layer4/udp/socket.rs
@@ -12,7 +12,7 @@ use crate::{
             PacketBuf,
             SharedLayer3Endpoint,
         },
-        udp::{
+        layer4::udp::{
             datagram::UdpDatagram,
             UdpHeader,
         },
